% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grids.R
\name{add_grid_neighbours}
\alias{add_grid_neighbours}
\title{Title}
\usage{
add_grid_neighbours(
  map_grid,
  direction = c("s-n", "n-s"),
  page_variable = "PageNumber",
  sep = "_",
  add_neighbours = TRUE
)
}
\arguments{
\item{add_neighbours}{}
}
\description{
Title
}
