% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_index.R
\name{add_index_1}
\alias{add_index_1}
\title{Title}
\usage{
add_index_1(
  .data,
  .grid_size,
  .fn,
  .prefix = "grid",
  .suffix = "",
  .keep = "all",
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
