% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_grids_rt90.R
\name{add_rt90_index}
\alias{add_rt90_index}
\title{Title}
\usage{
add_rt90_index(
  .data,
  .grid_size = c(50000, 25000, 5000),
  .prefix = "grid",
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
