% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_grids_sweref99.R
\name{add_sweref99_index}
\alias{add_sweref99_index}
\title{Title}
\usage{
add_sweref99_index(
  .data,
  .grid_size = c(1e+05, 50000, 25000, 10000, 5000, 2500, 1000, 500),
  crs = sf::st_crs(.data)$epsg,
  .prefix = "grid"
)
}
\arguments{
\item{.prefix}{}
}
\description{
Title
}
