% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrology.R
\name{aroid_group_dissolve}
\alias{aroid_group_dissolve}
\title{Title}
\usage{
aroid_group_dissolve(.x, .x_attr, .field = aroid_grp, .attr_field = aroid)
}
\arguments{
\item{.attr_field}{}
}
\description{
Title
}
