% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_wrappers.R
\name{copy_from_gpkg}
\alias{copy_from_gpkg}
\title{Title}
\usage{
copy_from_gpkg(
  dsn_source,
  dsn_destination,
  driver = c("GPKG", "OpenFileGDB"),
  filter_expr = NULL
)
}
\arguments{
\item{filter_expr}{}
}
\description{
Title
}
