% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_helper_functions.R
\name{count_points_in_polygons}
\alias{count_points_in_polygons}
\title{Title}
\usage{
count_points_in_polygons(polygons, points, fn = st_contains, name = "n_total")
}
\arguments{
\item{name}{}
}
\description{
Title
}
