% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates_transformation.R
\name{geodetic_to_grid}
\alias{geodetic_to_grid}
\title{Title}
\usage{
geodetic_to_grid(longitude, latitude, crs, colnames = c("y", "x"), p = NULL)
}
\arguments{
\item{p}{}
}
\description{
Title
}
