% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coord.R
\name{get_coord}
\alias{get_coord}
\title{Title}
\usage{
get_coord(
  lat,
  lon,
  simplify = FALSE,
  from_coord_sys = "AUTO",
  to_coord_sys = "4326",
  encoding = "UTF-8",
  karta = FALSE
)
}
\arguments{
\item{karta}{}
}
\description{
Title
}
