% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpx_to_sf.R
\name{gpx_to_sf}
\alias{gpx_to_sf}
\title{Title}
\usage{
gpx_to_sf(
  .x,
  data_type = c("route_points", "track_points", "waypoints"),
  bearing = FALSE,
  convert_time = TRUE,
  current_time_zone = TRUE,
  lines = FALSE,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
