% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grids.R
\name{grid_cell}
\alias{grid_cell}
\title{Title}
\usage{
grid_cell(
  x_min,
  y_min,
  delta_x,
  delta_y,
  direction = c("clockwise", "counter-clockwise")
)
}
\arguments{
\item{direction}{}
}
\description{
Title
}
