% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grids.R
\name{grid_parms}
\alias{grid_parms}
\title{Title}
\usage{
grid_parms(
  delta_x,
  delta_y,
  min_x = 2e+05,
  max_x = 1e+06,
  min_y = 6100000,
  max_y = 7700000
)
}
\arguments{
\item{max_y}{}
}
\description{
Title
}
