% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates_transformation.R
\name{grid_to_geodetic}
\alias{grid_to_geodetic}
\title{Title}
\usage{
grid_to_geodetic(y, x, crs, colnames = c("latitude", "longitude"), p = NULL)
}
\arguments{
\item{p}{}
}
\description{
Title
}
