% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates_transformation.R
\name{grid_to_grid}
\alias{grid_to_grid}
\title{Title}
\usage{
grid_to_grid(
  .data,
  y,
  x,
  crs_from,
  crs_to,
  round = FALSE,
  round_grid = 0,
  round_degrees = 6,
  colnames_new = c("y", "x")
)
}
\arguments{
\item{colnames_new}{}
}
\description{
Title
}
