% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpx_to_sf.R
\name{import_gpx}
\alias{import_gpx}
\title{Title}
\usage{
import_gpx(
  x,
  layer = "waypoints",
  .cols = c(name, cmt, geometry),
  crs = 3847,
  .new_cols = c("Easting90", "Northing90")
)
}
\arguments{
\item{.new_cols}{}
}
\description{
Title
}
