% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grids.R
\name{layout_grid_size}
\alias{layout_grid_size}
\title{Title}
\usage{
layout_grid_size(
  map_frame_x = 37.9,
  map_frame_y = 27.2,
  map_scale = 50000,
  overlap = 1.1,
  round = FALSE
)
}
\arguments{
\item{round}{}
}
\description{
Title
}
