% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet.R
\name{lm_basemaps}
\alias{lm_basemaps}
\title{Title}
\usage{
lm_basemaps(
  .data = NULL,
  crs = leaflet_crs_3006(),
  wms_layers = wms_layers_data,
  setView_lng = 16.15043,
  setView_lat = 59.60423,
  setView_zoom = 2,
  addOverlayGroups = NULL,
  collapseLayersControl = FALSE,
  add_extras = TRUE
)
}
\arguments{
\item{add_extras}{}
}
\description{
Title
}
