% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet.R
\name{lm_basemaps_old}
\alias{lm_basemaps_old}
\title{Title}
\usage{
lm_basemaps_old(
  .data = NULL,
  crs = leaflet_crs_3006(),
  topowebb_url = "https://minkarta.lantmateriet.se/map/topowebb/?",
  ortofoto_url = "https://minkarta.lantmateriet.se/map/ortofoto/?",
  hojdmodell_url = "https://minkarta.lantmateriet.se/map/hojdmodell/?",
  fastigheter_url = "https://minkarta.lantmateriet.se/map/fastighetsindelning/?",
  historiska_ortofoton_url =
    "https://api.lantmateriet.se/historiska-ortofoton/wms/v1/token/a2c113be5a3969e03dbd50222d3e2550/?",
  marktacke_url =
    "https://nvpub.vic-metria.nu/arcgis/services/NMD2018/MapServer/WMSServer",
  setView_lng = 16.15043,
  setView_lat = 59.60423,
  setView_zoom = 2,
  addOverlayGroups = NULL,
  CollapseLayerscontrol = FALSE
)
}
\arguments{
\item{CollapseLayerscontrol}{}
}
\description{
Title
}
