% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates_transformation_parameters.R
\name{proj_crs}
\alias{proj_crs}
\title{Title}
\usage{
proj_crs(proj_path = base::system.file("proj/proj.db", package = "sf"))
}
\arguments{
\item{proj_path}{}
}
\description{
Title
}
