% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates_transformation_proj.R
\name{proj_trans_add_cols}
\alias{proj_trans_add_cols}
\title{Title}
\usage{
proj_trans_add_cols(.data, coords, crs_from, crs_to, coords_names = NULL)
}
\arguments{
\item{coords_names}{}
}
\description{
Title
}
