% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates_transformation_proj.R
\name{proj_trans_many_crs}
\alias{proj_trans_many_crs}
\title{Title}
\usage{
proj_trans_many_crs(
  .data,
  coords = c("x", "y"),
  crs = "crs",
  crs_to = c(3006, 3847, 4619),
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
