% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_grids_rt90.R
\name{rt90_index}
\alias{rt90_index}
\title{Title}
\usage{
rt90_index(
  .x,
  .y = NULL,
  .grid_size = 5000,
  pad0 = FALSE,
  space = TRUE,
  caps = FALSE,
  rubin = FALSE,
  rubin_num = 1000,
  rubin_space = TRUE,
  y_min = 6100000,
  y_max = 7700000,
  x_min = 1200000,
  x_max = 1900000
)
}
\arguments{
\item{x_max}{}
}
\description{
Title
}
