% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_helper_functions.R
\name{st_add_geom_column_round}
\alias{st_add_geom_column_round}
\title{Title}
\usage{
st_add_geom_column_round(.x, .g, .g_names, centroid = TRUE, crs = 3006)
}
\arguments{
\item{crs}{}
}
\description{
Title
}
