% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_helper_functions.R
\name{st_nearest_feature_with_distance}
\alias{st_nearest_feature_with_distance}
\title{Title}
\usage{
st_nearest_feature_with_distance(x, y, output = c("x", "y"))
}
\arguments{
\item{output}{}
}
\description{
Title
}
