% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{storrutor}
\alias{storrutor}
\title{Swedish Land Survey Data}
\format{
\subsection{\code{storrutor}}{

A data frame with 252 rows and 5 columns, 50 x 50 km grid cells covering Sweden in RT90 2.5 gon V (EPSG:3021):
\describe{
\item{ruta_id}{Grid cell, first two numeric characters gives Northing position, last alphabetic character gives Easting Position}
\item{ruta}{Alternative representation of ruta_id, removes leading zero where present.}
\item{namn}{Geographic name of grid cell}
\item{northing, easting}{Geographic coordinates in EPSG:3021 of lower left corner of grid cell}
...
}
}
}
\source{
\subsection{Grid cells name were extracted from the Swedish Land Survey Data software Kartex, discontinued in 2015.}{
}
}
\usage{
storrutor
}
\description{
A subset of data from the Swedish Land Survey Data
Report ...
}
\keyword{datasets}
