% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates_transformation.R
\name{swe_transform}
\alias{swe_transform}
\title{Title}
\usage{
swe_transform(
  ost,
  nord,
  crs_from,
  crs_to = c(3006, 3847),
  colnames_geodetic = c("latitude", "longitude"),
  colnames_grid_prefix = c("north", "east"),
  colnames_grid = purrr::map(crs_to, function(x) str_c(colnames_grid_prefix, x, sep =
    "_")),
  round = FALSE,
  round_grid = 0,
  round_degrees = 6,
  keep_original = TRUE
)
}
\arguments{
\item{keep_original}{}
}
\description{
Title
}
