% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_grids_sweref99.R
\name{sweref99_index_alphanum}
\alias{sweref99_index_alphanum}
\title{Title}
\usage{
sweref99_index_alphanum(
  .x,
  .y = NULL,
  .grid_size,
  crs = 3006,
  caps = FALSE,
  space = TRUE,
  fastighetsblad = FALSE
)
}
\arguments{
\item{fastighetsblad}{}
}
\description{
Title
}
