% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet.R
\name{wms_add}
\alias{wms_add}
\title{Title}
\usage{
wms_add(
  map,
  group,
  url,
  layers,
  format = "image/png",
  transparent = TRUE,
  opacity = 1,
  styles = "",
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
