% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet.R
\name{wms_sources}
\alias{wms_sources}
\title{Title}
\usage{
wms_sources(
  path = "inst/extdata/wms_sources.xlsx",
  layers_sheet = "wms_layers",
  url_sheet = "wms_url",
  sep = ", "
)
}
\arguments{
\item{sep}{}
}
\description{
Title
}
